Add files via upload
As of now, the program reads one .csv file of NIFTYBANK, creates a candle of first 15 minutes and initiate the trade based on the given conditions with a stop loss of 0.5%.. 
Further it checks whether the stock has hit the stop loss or not and does the required operations as mentioned in the task.  

This is just a basic work flow of the project. there still needs a lots of modifications.